+++ work_1/src/TodoListItem.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Todo_ListItem = ({label}) => {
+    return (
+        <span>{label}</span>
+    )
+}
+
+export default Todo_ListItem
