+++ work_1/src/TodoSearch.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Todo_Search = () => {
+    return (
+        <div>
+            <input type="text"/>
+        </div>
+    )
+}
+
+export default Todo_Search
