+++ work_1/src/TodoStatusFilter.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+const Todo_StatusFilter = () => {
+    return (
+        <div>
+            <button>All</button>
+            <button>Active</button>
+            <button>Dote</button>
+        </div>
+    )
+}
+
+export default Todo_StatusFilter
